MSDK-315: Standardize error handling and eliminate silent failures (#256)
* MSDK-315: Standardize error handling and eliminate silent failures

- Consolidate ATTNSDKError into ATTNError with new cases (initializationFailed,
  missingPushToken, httpError) and deprecate ATTNSDKError
- Add CustomNSError conformance for Obj-C consumers (error domain, codes, userInfo)
- Fix 4 silent failure paths in ATTNAPI where callbacks were never invoked
- Fix HTTP 400 detection bug (statusCode > 400 missed 400 itself)
- Standardize all HTTP status checks via HTTPURLResponse.isSuccessful extension
- Add update(domain:completion:) overload for proper error propagation
- Add domain validation to static ATTNSDK.initialize() method
- Add ATTNAPISpy.stubbedError for testing error propagation
- Add comprehensive unit tests for error types and propagation


* Address Codex review: fix threading contract and preserve legacy error types

1. Wrap invalid-domain early-return completion in DispatchQueue.main.async
   to match the threading contract of all other completion paths.

2. Restore ATTNSDKError emission in existing callback paths to preserve
   backward compatibility for clients pattern-matching on the legacy type.

3. Add "Backward Compatibility" guidelines to CLAUDE.md documenting that
   breaking changes must be avoided — deprecate, don't remove.


* Address PR review: remove unused ATTNError cases, fix update(domain:) threading

1. Remove ATTNError.missingPushToken and .httpError — no call site emits
   these yet. They'll be re-added with MSDK-298 Phase 3 when wired up,
   avoiding shipping dead public API that's costly to remove later.

2. Dispatch all update(domain:completion:) callbacks via
   DispatchQueue.main.async to match the threading contract established
   by static initialize(domain:mode:completion:).

3. Update tests to use XCTestExpectation for the now-async completions.


* Extract marketing subscriptions into ATTNSDK+MarketingSubscriptions.swift

Fixes SwiftLint type_body_length error (707 lines, limit 700) caused by
MSDK-344 pushEnabled additions combined with MSDK-315 error handling.


---------
