+++ Sources/API/ATTNError.swift
@@ -7,13 +7,14 @@
 
 import Foundation
 
-public enum ATTNError: Error {
+public enum ATTNError: Error, Equatable {
     case sdkNotInitialized
     case missingContactInfo
     @available(*, deprecated, message: "Geo-domain adjustment has been removed. This case is no longer used by the SDK.")
     case geoDomainUnavailable
     case badURL
     case invalidDomain
+    case initializationFailed
 }
 
 extension ATTNError: LocalizedError {
@@ -29,6 +30,23 @@ extension ATTNError: LocalizedError {
             return "Invalid URL"
         case .invalidDomain:
             return "The provided domain is not recognized. Please verify that the domain matches your Attentive settings."
+        case .initializationFailed:
+            return "SDK initialization failed"
+        }
+    }
+}
+
+extension ATTNError: CustomNSError {
+    public static var errorDomain: String { "com.attentive.sdk" }
+
+    public var errorCode: Int {
+        switch self {
+        case .sdkNotInitialized: return 1
+        case .missingContactInfo: return 2
+        case .geoDomainUnavailable: return 3
+        case .badURL: return 4
+        case .invalidDomain: return 5
+        case .initializationFailed: return 6
         }
     }
 }
