+++ Sources/Helpers/Extension/HTTPURLResponse+Extension.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+extension HTTPURLResponse {
+    var isSuccessful: Bool { (200...299).contains(statusCode) }
+}
